# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
}

export interface AuthResponse {
  id: number;
  name: string;
  email: string;
  token: string;
}

export interface RegisterInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type Category = typeof Category[keyof typeof Category];


export const Category = {
  Makanan: 'Makanan',
  Transportasi: 'Transportasi',
  Belanja: 'Belanja',
  Hiburan: 'Hiburan',
  Kesehatan: 'Kesehatan',
  Tagihan: 'Tagihan',
  Lainnya: 'Lainnya',
} as const;

export interface Transaction {
  id: number;
  title: string;
  category: Category;
  amount: number;
  /** @nullable */
  note?: string | null;
  occurredAt: string;
}

export interface TransactionInput {
  /** @minLength 1 */
  title: string;
  category: Category;
  /** @minimum 1 */
  amount: number;
  note?: string;
}