# lib/api-zod/src/generated/types/authResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthResponse {
  id: number;
  name: string;
  email: string;
  token: string;
}